torch.ao migration: stubs.py phase 1
1. move the file
  ```
  hg mv caffe2/torch/quantization/stubs.py caffe2/torch/ao/quantization/
  ```

  2. create a new file in the old location and copy the imports
  3. fix all call sites inside `torch`

Differential Revision: [D30879678](https://our.internmc.facebook.com/intern/diff/D30879678/)

[ghstack-poisoned]
