+++ test/quantization/ao_migration/test_quantize.py
@@ -61,3 +61,14 @@ class TestAOMigrationQuantizePy(AOMigrationTestCase):
             'swap_module',
         ]
         self._test_function_import('quantize', function_list)
+
+    def test_package_import_stubs(self):
+        self._test_package_import('stubs')
+
+    def test_function_import_stubs(self):
+        function_list = [
+            'QuantStub',
+            'DeQuantStub',
+            'QuantWrapper',
+        ]
+        self._test_function_import('stubs', function_list)
